requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch
torchvision
torchaudio
pandas
Pillow
scikit-learn
transformers
tqdm
tensorboard